[KFKESQ-11] Removed the niu package and its classes
