--- kafkaesque-core/src/main/java/in/rcardin/kafkaesque/niu/KafkaesqueInputTopicNew.java
@@ -1,21 +1,0 @@
-package in.rcardin.kafkaesque.niu;
-
-import in.rcardin.kafkaesque.niu.KafkaesqueProducerNew.Builder;
-import java.util.List;
-
-public class KafkaesqueInputTopicNew<Key, Value> {
-
-  private final Builder<Key, Value> producerBuilder;
-  
-  KafkaesqueInputTopicNew(
-      Builder<Key, Value> producerBuilder) {
-    this.producerBuilder = producerBuilder;
-  }
-  
-  public void pipeRecordList(List<KafkaesqueProducerNew.Record<Key, Value>> records) {
-    producerBuilder
-        .messages(records)
-        .expecting()
-        .assertingAfterAll(producerRecords -> {});
-  }
-}
