+++ kafkaesque-core/src/test/java/in/rcard/kafkaesque/KafkaesqueConsumerTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.BDDMockito.given;
 
-import in.rcard.kafkaesque.KafkaesqueConsumer.KafkaesqueConsumerDelegate;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
