+++ kafkaesque-core/src/test/java/in/rcard/kafkaesque/KafkaesqueTest.java
@@ -8,8 +8,6 @@ import static org.mockito.BDDMockito.anyString;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 
-import in.rcard.kafkaesque.KafkaesqueConsumer.Builder;
-import java.util.concurrent.TimeUnit;
 import org.apache.kafka.common.serialization.IntegerDeserializer;
 import org.apache.kafka.common.serialization.IntegerSerializer;
 import org.apache.kafka.common.serialization.StringDeserializer;
