+++ kafkaesque-live-kafka/src/test/java/in/rcard/kafkaesque/LiveKafkaesqueTest.java
@@ -2,8 +2,6 @@ package in.rcard.kafkaesque;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import in.rcard.kafkaesque.KafkaesqueOutputTopic.Message;
-import in.rcard.kafkaesque.KafkaesqueProducer.Record;
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.Collection;
