+++ kafkaesque-spring-kafka/src/main/java/in/rcard/kafkaesque/SpringKafkaesque.java
@@ -1,7 +1,5 @@
 package in.rcard.kafkaesque;
 
-import in.rcard.kafkaesque.KafkaesqueConsumer.Builder;
-import in.rcard.kafkaesque.KafkaesqueConsumer.KafkaesqueConsumerDelegate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
